requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

